local_app/requirements.txt
streamlit>=1.28.0
strands
boto3
python-dotenv
tavily-python
bedrock-agentcore